README: drop TODO for --itemize-changes
The original idea with noting this as a TODO was that we could
try to generate output in the same format as rsync. However, I'm
not convinced it makes sense or is feasible due to the differences
in the backend. For example, unlike rsync, we have no idea whether
or not our publish is updating an existing URI.