Introduce diagnostic mode
This adds a new option which will dump a bunch of debugging info
when the command starts. This will log info which would otherwise
not be readily accessible in the normal course of events, e.g.:

- the content of the source tree is dumped in unprocessed
  form (e.g. show links as they are without following them)

- the content of the --files-from file is dumped

- exodus-gw is sanity checked and the current user/roles are logged

The primary motivation is that it is currently difficult to debug when
something goes wrong during an rhsm-pulp publish, because pulp does all
its work under a temporary directory which is immediately deleted at the
end of publish (whether it succeeds or fails). This means we normally
can't see exactly what was in the source tree or filters by the time we
know that a task has failed.

Enabling diagnostic mode can help by dumping all needed info into logs
where it can be inspected later, even after the source tree has been
deleted.