logging: increase verbosity when adding publish items
As far as I can tell, we have not previously logged the precise
items we're sending to exodus-gw even in debug mode, only logging
a summary for each batch.

Let's make it more verbose and produce debug logs for every
single item, as otherwise it's difficult to troubleshoot issues
with the calculation of each item (e.g. setting of path).