Generate link_to for symlinks if --links [RHELDST-8588]
When running in exodus mode, if the --links (-l) flag is set, any symlinks
will be copied rather than followed. This means that instead of resolving
the symlink and generating an object_key (sha256sum), link_to will be set
to the link's target relative to the rsync destination.

Ex.:
exodus-rsync --links /src exodus:/dest
body = [
    {
        "web_uri": "/dest/my/awesome/file.iso",
        "object_key": "aec070645fe53ee3b3763059376134f058cc337247c978add178b6ccdfb0019f",
    },
        "web_uri": "/dest/another/route/to/my/awsome/file.iso",
        "link_to": "/dest/my/awesome/file.iso",
    },
]