Fix calculation of link_to values [RHELDST-8935]
Prior to this commit, when attempting to publish items in the
"src" tree shown below to some destination (e.g., "exodus:/dest"):

$ tree src
src
└── some
    ├── dir
    │   └── link-to-somefile -> ../somefile
    └── somefile

2 directories, 2 files

The publish would generate two items:
1. {uri="/dest/some/somefile",
    object_key=$(somefile's checksum)}
2. {uri="/dest/some/dir/link-to-somefile",
    link_to="/dest/somefile"}

The publish could not be committed since no item exists with a uri
matching that value of link_to.

Now, the publish generates two items:

1. {uri="/dest/some/somefile",
    object_key=$(somefile's checksum)}
2. {uri="/dest/some/dir/link-to-somefile",
    link_to="/dest/some/somefile"}

The publish can be committed successfully.