Fix request error responses [RHEL-9209]
Previously, when an exodus-gw response status was between 200 and 299,
exodus-rsync would attempt to read the response body and include it in
the error message. The body read was incomplete (truncated) and
included lots of NULL bytes.

This commit changes how the response body is read to get the entire
message without including extra space filled with NULL bytes.

Signed-off-by: Nathan Gillett <ngillett@redhat.com>