Add a Makefile recipe for building in a container image
'make' by default is going to build exodus-rsync using a locally
installed go toolchain. That remains the default, but let's also
make it easy to build inside of a ubi8/go-toolset container image.

The idea is that this may be more appropriate for some CI environments
where we have the ability to run containers and don't have the go
toolchain already installed.