Refactor upload to parallelize existence checks [RHELDST-13868]
Commit 91a0db4b2b8345a introduced goroutines to the upload process so
that multiple files could be uploaded at once.

However, the part of the process checking whether a file already exists
in the bucket was still not parallelized. According to testing as seen
in RHELDST-13868, for a large push it is possible even for the existence
checks to take several hours, so it makes sense to support parallelizing
this as well.

This commit refactors the upload step so the existence checks also
happen concurrently. In local testing with default config this reduces
the time taken to check 1000 blobs from 60 to 15 seconds.

Making the existence checks concurrent is the primary motivation of this
refactor, but it is also improved in two other ways:

- callbacks are now all invoked as the uploads are still ongoing.
  Previously the onUploaded callback would only be invoked after *all*
  uploads had already completed, which would make it of limited use for
  reporting progress (possibly impacting RHELDST-13869).

- uploads are now cancelled as soon as a single error occurs, rather
  than continuing to process all other items even after we already know
  we're going to fail.

The refactored code remains 100% covered by existing tests.