Update to Go 1.19 [RHELDST-14916]
Bump the version of Go used in CI and for building releases from 1.16 to
1.19. The primary motivations are general maintenance, and unblocking
some dependabot updates which appear to require Go > 1.16.

One backwards-incompatible change required updates to a few tests: per
https://tip.golang.org/doc/go1.19#os-exec-path ,
PATH lookup relative to the current directory is no longer allowed for
security reasons. Some tests relied on this, so had to be refactored to
use absolute paths instead.

Other changes triggered by the upgrade include:

- in go.mod, indirect deps are now listed in a separate block
- "go fmt" made some minor whitespace changes