Support different commit modes [RHELDST-20490]
This adds support to exodus-rsync for using the new commit_mode argument
added to the exodus-gw commit API in [1]. The use-case is to allow
rhsm-pulp to do a phase1 commit during Pulp publish tasks, to ensure
that all RPMs have reached the CDN storage and later fast-forward
publishes will work correctly even if the current exodus-gw publish
later fails.

The default behavior is not to pass any `commit_mode`, which ensures
that exodus-rsync remains compatible with older versions of exodus-gw.

This change is aiming for flexibility:

- the argument's value is intentionally not validated as being "phase1"
  or "phase2"; validation is left to the server. This avoids
  exodus-rsync needing further code changes if more commit modes become
  supported later.

- the mode can be configured both using command-line arguments and the
  config file. The intent is to use the command-line argument from
  pubtools-exodus, but it seems wise to also support using the config
  file for cases not yet anticipated.

[1] https://github.com/release-engineering/exodus-gw/pull/596