Implement retries at http RoundTripper level [RHELDST-22079]
While net/http automatically retries on some types of error, it
doesn't seem to be enough to cope with certain issues in practice.
Add retry logic at the RoundTripper level using a library designed for
that.

The retry setup here is close to what the S3 client is already doing
internally, so this should cause requests to the gw publish API to
be handled more consistently with requests to the upload API.