Bump the default value of `gwmaxattempts` [RHELDST-22390]
Inspired by ticket RHELDST-22390 where a temporary issue in exodus-gw
resulted in a push failure. Based on the logs, retries only occurred for
a total of ~5 seconds or so which was not enough to recover from the
issue. The service did self-heal and an increased retry count probably
would have allowed exodus-rsync to ultimately succeed.