S3 client max retires as num of gwmaxattempts from config file
S3 client retires by default for 3 times rather than applying
the retires(gwmaxattempts) defined in the config file. It uses
a separate HTTP client than rest of the HTTP requests. So, the
config wasn't applicable to the s3 requests.
Hence, this updates the s3 client to apply the retry count
defined in the config file for all s3 requests as well.