Update to Go 1.23 [RHELDST-26109]
Bump the version of Go used in CI and for building releases from 1.19 to 1.23. This brings us onto the latest stable toolchain - a point of maintenance - and unblocks update of golang.org/x/crypto which requires Go >= 1.20.

Go 1.20 introduced a dynamic linker for glibc in work to make builds deterministically reproducible.
Staying glibc-compatible with RHEL 6 now forces use of a Debian 11 platform with older glibc: container image golang:1.23-bullseye and GitHub Runner Ubuntu 20.04.