Address duplicate uploads [RHELDST-30656]
Previously, uploadWorkers only had HEAD requests to the exodus S3
bucket to determine whether or not an object needed to be uploaded.

This allowed for the possibility for multiple workers to upload the same
object, should they operate on the object in a close enough timeframe.

This commit implements a concurrency-safe map, to track which objects
have begun the upload process, and a check in the worker to prevent
operating on an object already recorded in that map.